Update README.md
Write a Java code for each problem. Make sure your program satisfies all specified requirements. When an example result is given, output of your program should match the result.

Problem 11 (20 points)

A palindrome is a word, number, phrase, or other sequence of characters which reads the same backward as forward, such as madam and racecar (from wikipedia). In this problem, you will write a program that checks whether a given string or integer is a palindrome or not.

You are given the code for class Problem11, which must NOT be modified. Write other parts of the code to complete the program.

public class Problem11 {

public static void main(String[] args) {